+++ glancy-site/src/api/client.js
@@ -1,4 +1,4 @@
-import { extractMessage } from '@/utils/index.js'
+import { extractMessage } from '@/utils/json.js'
 
 /**
  * Create a new API client instance with optional default headers and token.
