+++ glancy-site/src/store/historyStore.ts
@@ -1,6 +1,6 @@
 import { create } from 'zustand'
 import api from '@/api/index.js'
-import { safeJSONParse } from '@/utils/index.js'
+import { safeJSONParse } from '@/utils/json.js'
 
 import type { User } from './userStore.ts'
