+++ glancy-site/src/store/userStore.ts
@@ -1,5 +1,5 @@
 import { create } from 'zustand'
-import { safeJSONParse } from '@/utils/index.js'
+import { safeJSONParse } from '@/utils/json.js'
 
 export interface User {
   id: string
