+++ glancy-site/src/utils/json.js
@@ -1,0 +1,20 @@
+export function extractMessage(text) {
+  if (!text) return ''
+  try {
+    const data = JSON.parse(text)
+    if (data && typeof data === 'object') {
+      return data.message || text
+    }
+  } catch {
+    // not JSON, ignore
+  }
+  return text
+}
+
+export function safeJSONParse(str, defaultValue = null) {
+  try {
+    return JSON.parse(str)
+  } catch {
+    return defaultValue
+  }
+}
